// apps/pwabuilder-chromium-extension/src/interfaces/manifest.ts
export interface Icon {
  src: string;
  generated?: boolean;
  type?: string;
  sizes?: string;
  purpose?: "any" | "maskable" | "monochrome";
  label?: string;
}